Fix gnosis icon missing
