+++ shared/src/assets/icons/defiApp.tsx
@@ -1,0 +1,67 @@
+import React from "react";
+
+type SVGProps = React.SVGAttributes<SVGElement>;
+
+export const YearnLogo: React.FC<SVGProps> = (props) => (
+  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1000 1000" {...props}>
+    <title>yearn-finance-yfi</title>
+    <g id="Layer_2" data-name="Layer 2">
+      <g id="Layer_1-2" data-name="Layer 1">
+        <path
+          id="SVGID"
+          className="background"
+          d="M500,0c276.1,0,500,223.9,500,500s-223.9,500-500,500S0,776.1,0,500,223.9,0,500,0Z"
+          fill="#006ae3"
+          fillRule="evenodd"
+        />
+        <polygon
+          id="SVGID-2"
+          className="content"
+          data-name="SVGID"
+          points="479.4 710.8 479.4 294.8 524.7 294.8 524.7 710.8 479.4 710.8"
+          fill="#fff"
+        />
+        <path
+          id="SVGID-3"
+          className="content"
+          data-name="SVGID"
+          d="M710,433.4,570.1,470.5,538.9,325.3l41.6-9.4,16.4,68.8s37.8-62-12.6-126.2c-29.7-33-43.8-34.4-77.1-39.6-29.3-4.2-97.4,5.7-117.7,85.2-8.6,51.2,1.1,89.1,67.1,138.7L452.9,498s-73.7-51.9-92.7-88.3c-14.7-28.8-39.9-85.7,5.6-164.9,24.5-39.6,72.8-77.6,157.9-73.6C566.5,173,671,225.3,654.8,347.5c-2.8,22.9-14.7,53.4-14.7,53.4l57.4-12.8L710,433.4Z"
+          fill="#fff"
+        />
+        <path
+          id="SVGID-4"
+          className="content"
+          data-name="SVGID"
+          d="M630.1,758.9c-25.6,38.9-74.9,75.6-159.8,69.4-42.8-2.9-145.8-58-126.4-179.7,3.4-22.8,16.1-52.9,16.1-52.9l-57.7,11.2L291,561.4,431.9,528l27.4,146-41.9,8.3L402.8,613s-39.4,60.9,9.2,126.5c28.8,33.8,42.8,35.5,76.1,41.6,29.1,5,97.5-3.2,119.9-82.1,9.9-50.9,1.2-89.1-63.4-140.5l5.2-55.1s72.3,53.8,90.3,90.7c13.8,29.3,37.5,86.9-10,164.8Z"
+          fill="#fff"
+        />
+      </g>
+    </g>
+  </svg>
+);
+
+export const GnosisLogo: React.FC<SVGProps> = (props) => (
+  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 2000 2000" {...props}>
+    <g fill="#00a6c4">
+      <path d="M1809.4 500l-73.6 73.6c58.5 92 76.9 207.4 36.8 317.7-66.9 187.3-274.2 286-463.2 219.1-26.8-10-51.8-21.7-73.6-36.8L1000 1309.4 794.3 1102c-95.3 63.5-217.4 81.9-332.8 41.8-194-70.2-294.3-282.6-224.1-474.9 10-30.1 25.1-56.9 41.8-81.9l-88.6-88.6-16.7 28.4c-92 150.5-142.1 324.4-142.1 503.3-1.7 533.4 433.1 969.9 966.5 969.9h1.7c533.4 0 966.6-433.1 968.2-966.6 0-177.3-48.5-351.2-140.5-503.3l-18.3-30.1" />
+      <path d="M388 695.7c-25.1 33.4-40.1 75.3-40.1 120.4 0 108.7 88.6 197.3 197.3 197.3 45.2 0 87-15.1 120.4-41.8L388 695.7M1356.2 954.8c31.8 21.7 68.6 33.4 110.4 33.4 108.7 0 197.3-88.6 197.3-197.3 0-40.1-11.7-78.6-33.4-110.4l-274.3 274.3M1003.3 1162.2L204 359.5l31.8-33.4C433.1 117.1 702.3 0 991.6 0h1.7c292.6 0 573.6 125.4 769.2 342.8l30.1 33.4-789.3 786M334.5 359.5l668.9 670.6 660.5-657.2C1486.6 195.7 1245.8 93.6 995 93.6h-1.7c-249.2 0-481.6 93.7-658.8 265.9" />
+    </g>
+  </svg>
+);
+
+export const FlexLogo: React.FC<SVGProps> = (props) => (
+  <svg
+    width="27"
+    height="24"
+    viewBox="0 0 27 24"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path d="M0 24V0H3.15328V8.05839H10.6861L0 24Z" fill="#14F195" />
+    <path
+      d="M26.2774 24H22.9489V16.1168H15.2409L26.2774 0V24Z"
+      fill="#14F195"
+    />
+  </svg>
+);
